% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{as_clever_tbl}
\alias{as_clever_tbl}
\title{Coerce a data.frame into a clever_tbl.}
\usage{
as_clever_tbl(x)
}
\arguments{
\item{x}{a data.frame}
}
\value{
a clever_tbl
}
\description{
as_clever_tbl turns a data.frame into a clever_tbl, a data frame with class clever_tbl.
}
\examples{
mtcars \%>\% as_clever_tbl()
}
