% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/askia_cleaners.R
\name{askia_cleaners}
\alias{askia_cleaners}
\alias{askia_clean_label}
\alias{askia_clean_levels}
\alias{askia_as_integer}
\alias{askia_as_ordered}
\alias{askia_as_reverse_ordered}
\alias{askia_as_likert}
\alias{askia_reorder_multi}
\alias{askia_replace_multi}
\title{Cleanup Askia output}
\usage{
askia_clean_label(x)

askia_clean_levels(x)

askia_as_integer(x)

askia_as_ordered(x, ...)

askia_as_reverse_ordered(x)

askia_as_likert(x)

askia_reorder_multi(df, pattern, label = NULL)

askia_replace_multi(df, df2)
}
\arguments{
\item{x}{a variable or data.frame}

\item{...}{further arguments passed to or from other methods}

\item{df}{a data.frame}

\item{pattern}{pattern in the names of the variables that form a 'multi'}

\item{label}{label for the 'multi' (often the question that was asked)}

\item{df2}{a second data.frame}
}
\value{
the cleaned variable(s)
}
\description{
Cleanup Askia output
}
\details{
\code{askia_clean_label} removes html from a variable's label attribute

\code{askia_clean_levels} removes html from factor levels

\code{askia_as_integer} coerces a variable to integer while keeping the label attribute

\code{askia_as_ordered} coerces a variable to an ordered factor while keeping the label attribute

\code{askia_as_reverse_ordered} coerces a variable to a reverse-ordered factor while keeping the label attribute

\code{askia_as_likert} coerces a factor to a likert (an integer) if possible (i.e. if all but the highest and lowest levels are integer values).

\code{askia_reorder_multi} converts 'multi' variables so that they indicate presence instead of ordering of the alternatives

\code{askia_replace_multi} replaces 'multi' variables in \code{df} with their recoded version in \code{df2}
}
