% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleverR.R
\docType{package}
\name{cleverR}
\alias{cleverR}
\alias{cleverR-package}
\title{cleverR: Functions For Clever Notebooks.}
\description{
The foo package provides three categories of important functions:
foo, bar and baz.
}
\section{cleverR functions}{

The cleverR functions ...
}

