% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_summaries.R
\name{column_func}
\alias{column_func}
\title{Compute a property or univariate statistic for all variables (columns) in a dataframe}
\usage{
column_func(x, fun, newname = "descriptive")
}
\arguments{
\item{x}{a tibble}

\item{fun}{the function that computes the property or statistic}

\item{newname}{the name of the variable that will hold the result. Defaults to 'descriptive'.}
}
\value{
a tibble with the names of the variables in x (variable) and the statistic that was computed (newname)
}
\description{
Compute a property or univariate statistic for all variables (columns) in a dataframe
}
\keyword{internal}
