% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_summaries.R
\name{compute_descriptives}
\alias{compute_descriptives}
\alias{compute_descriptives_numeric}
\alias{compute_descriptives_logical}
\alias{compute_descriptives_ordered}
\alias{compute_descriptives_date}
\title{Compute a set of univariate statistics for all variables (columns) in a dataframe}
\usage{
compute_descriptives(x)

compute_descriptives_numeric(x)

compute_descriptives_logical(x)

compute_descriptives_ordered(x)

compute_descriptives_date(x)
}
\arguments{
\item{x}{a tibble}
}
\value{
a tibble with variables in the rows and univariate statistics in the columns
}
\description{
Compute a set of univariate statistics for all variables (columns) in a dataframe
}
\details{
\code{compute_descriptives} computes the statistics for numeric, logical and ordered variables

\code{compute_descriptives_numeric} computes the statistics for numeric variables

\code{compute_descriptives_logical} computes the statistics for logical variables

\code{compute_descriptives_ordered} computes the statistics for ordered variables

\code{compute_descriptives_date} computes the statistics for date variables
}
\examples{
datasets::mtcars \%>\% compute_descriptives()
}
