% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_summaries.R
\name{compute_size}
\alias{compute_size}
\title{Compute the size of a data.frame}
\usage{
compute_size(x)
}
\arguments{
\item{x}{a data.frame}
}
\value{
a tibble with the number of observations and the number of variables (statistic - value)
}
\description{
Compute the size of a data.frame
}
\examples{
datasets::mtcars \%>\% compute_size()
}
