% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_summaries.R
\name{compute_tables}
\alias{compute_tables}
\alias{compute_tables_factor}
\alias{compute_tables_character}
\alias{compute_tables_logical}
\alias{compute_tables_likert}
\title{Compute a set of frequency tables for all variables (columns) in a dataframe}
\usage{
compute_tables(x, max_distinct_values = 11)

compute_tables_factor(x, max_distinct_values = 11)

compute_tables_character(x, max_distinct_values = 11)

compute_tables_logical(x)

compute_tables_likert(x)
}
\arguments{
\item{x}{a tibble}

\item{max_distinct_values}{maximum number of entries in the table. Defaults to 11.}
}
\value{
a tibble with variables in the rows and univariate statistics in the columns
}
\description{
Compute a set of frequency tables for all variables (columns) in a dataframe
}
\details{
\code{compute_tables} computes the tables for numeric, logical and ordered variables

\code{compute_tables_factor} computes the tables for factors

\code{compute_tables_character} computes the tables for character variables

\code{compute_tables_logical} computes the tables for logical variables

\code{compute_tables_likert} computes the tables for Likert scales
}
\examples{
datasets::iris \%>\% compute_tables()
}
