% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_summaries.R
\name{descriptives}
\alias{descriptives}
\alias{mean_func}
\alias{sd_func}
\alias{quantile_func}
\alias{min_func}
\alias{Q1_func}
\alias{median_func}
\alias{Q3_func}
\alias{max_func}
\alias{mean_date_func}
\alias{sd_date_func}
\alias{quantile_date_func}
\alias{min_date_func}
\alias{Q1_date_func}
\alias{median_date_func}
\alias{Q3_date_func}
\alias{max_date_func}
\title{Helper functions that compute univariate summaries of a variable}
\usage{
mean_func(x)

sd_func(x)

quantile_func(x, prob)

min_func(x)

Q1_func(x)

median_func(x)

Q3_func(x)

max_func(x)

mean_date_func(x)

sd_date_func(x)

quantile_date_func(x, prob)

min_date_func(x)

Q1_date_func(x)

median_date_func(x)

Q3_date_func(x)

max_date_func(x)
}
\arguments{
\item{x}{a variable}

\item{prob}{probability in [0,1].}
}
\description{
Helper functions that compute univariate summaries of a variable
}
\details{
\code{mean_func} and \code{mean_date_func} return the mean of a (date) variable

\code{sd_func} and \code{sd_date_func} return the standard deviation of a (date) variable

\code{quantile_func} and \code{quantile_date_func} return a quantile of a (date) variable

\code{min_func} and \code{min_date_func} return the minimum of (date) variable

\code{Q1_func} and \code{Q1_date_func} return the first quartile of (date) variable

\code{median_func} and \code{median_date_func} return median of (date) variable
\code{Q3_func} and \code{Q3_date_func} return third quartile of (date) variable

\code{max_func} and \code{max_date_func} return maximum of (date) variable
}
\keyword{internal}
\keyword{internalt}
