% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate_summaries.R
\name{do_downsample}
\alias{do_downsample}
\title{Downsample a dataframe}
\usage{
do_downsample(x, downsample = "auto")
}
\arguments{
\item{x}{a data.frame}

\item{downsample}{downsampling factor. Defaults to "auto", in this case it keeps a maximum of 10000 observations.}
}
\value{
the downsampled data.frame
}
\description{
Reduces the number of observations in a data.frame for plotting.
}
\keyword{internal}
