% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate_summaries.R
\name{markdown_univariate}
\alias{markdown_univariate}
\title{Generate rmarkdown to plot all variables in a dataframe}
\usage{
markdown_univariate(x)
}
\arguments{
\item{x}{a data.frame}
}
\value{
prints out markdown on the console.
}
\description{
Generate rmarkdown to plot all variables in a dataframe
}
