% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate_summaries.R
\name{missing_group_plot}
\alias{missing_group_plot}
\title{Grouped missingness plot}
\usage{
missing_group_plot(x, downsample = "auto")
}
\arguments{
\item{x}{a data.frame}

\item{downsample}{downsampling factor. Defaults to "auto", in this case it keeps a maximum of 10000 observations.}
}
\description{
Grouped missingness plot
}
