% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_summaries.R
\name{name_func}
\alias{name_func}
\title{get the names of the variables in a data.frame}
\usage{
name_func(x)
}
\arguments{
\item{x}{a data.frame}
}
\value{
a tibble with the names of the variables in x (variable)
}
\description{
get the names of the variables in a data.frame
}
\keyword{internal}
