% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate_summaries.R
\name{panel_functions}
\alias{panel_functions}
\alias{panel_densityplot}
\alias{panel_boxplot}
\alias{panel_qqplot}
\alias{panel_barplot}
\alias{panel_missingplot}
\title{Panel functions for plotting variables}
\usage{
panel_densityplot(x, variable)

panel_boxplot(x, variable, downsample = "auto")

panel_qqplot(x, variable, downsample = "auto")

panel_barplot(x, variable)

panel_missingplot(x, variable, downsample = "auto")
}
\arguments{
\item{x}{a data.frame}

\item{variable}{the variable that is plotted}

\item{downsample}{downsampling factor. Defaults to "auto", in this case it keeps a maximum of 10000 observations.}
}
\description{
Panel functions for plotting variables
}
\details{
\code{panel_densityplot} generates a density plot

\code{panel_boxplot} generates a box plot

\code{panel_qqplot} generates a quantile-quantile plot against the quantiles of the normal distribution

\code{panel_barplot} generates a bar plot of the variable

\code{panel_missingplot} generates a missingness plot
}
\keyword{internal}
