% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate_summaries.R
\name{plot_univariate}
\alias{plot_univariate}
\alias{plot_continuous}
\alias{plot_likert}
\alias{plot_posixt}
\alias{plot_factor}
\alias{plot_string}
\alias{plot_logical}
\title{Summary plots for variables}
\usage{
plot_continuous(x, variable, downsample = "auto")

plot_likert(x, variable, downsample = "auto")

plot_posixt(x, variable, downsample = "auto")

plot_factor(x, variable, max_distinct_values = 11, downsample = "auto")

plot_string(x, variable, max_distinct_values = 11, downsample = "auto")

plot_logical(x, variable, downsample = "auto")
}
\arguments{
\item{x}{a data.frame}

\item{variable}{the variable that is plotted}

\item{downsample}{downsampling factor. Defaults to "auto", in this case it keeps a maximum of 10000 observations}

\item{max_distinct_values}{the maximum number of different levels or values in a factor or string variable}
}
\description{
Summary plots for variables
}
\details{
\code{plot_continuous} generates plots for continuous variables: a density plot, a box plot, a qq plot and a missingness plot

\code{plot_likert} generates plots for likert variables. Similar to a continuous plot, but the density plot is replaced by a bar plot.

\code{plot_posixt} generates plots for times. The same as the continuous plot but without the summary statistics

\code{plot_factor} generates plots for factor variables: a bar plot and a missingness plot

\code{plot_string} generates plots for string variables. The string is converted to a factor with \code{max_distinct_values} levels, then the factor plots are applied.

\code{plot_logical} generates plots for logical variables. The string is converted to a factor with two levels, then the factor plots are applied.
}
