% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{print.clever_tbl}
\alias{print.clever_tbl}
\title{Print a clever_tbl object}
\usage{
\method{print}{clever_tbl}(x, maxwidth = 80, ...)
}
\arguments{
\item{x}{a clever_tbl}

\item{maxwidth}{maximum width of a field. Defaults to 80}

\item{...}{further arguments passed to or from other methods}
}
\value{
returns the clever_tbl invisibly
}
\description{
Print a clever_tbl object
}
\examples{
mtcars \%>\% as_clever_tbl() \%>\% print()
}
