% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_summaries.R
\name{properties}
\alias{properties}
\alias{class_func}
\alias{levels_func}
\alias{labels_func}
\alias{observed_func}
\alias{missing_func}
\alias{distinct_func}
\title{Helper functions that compute properties of a variable}
\usage{
class_func(x)

levels_func(x)

labels_func(x)

observed_func(x)

missing_func(x)

distinct_func(x)
}
\arguments{
\item{x}{a variable}
}
\description{
Helper functions that compute properties of a variable
}
\details{
\code{class_func} returns the class of a variable

\code{levels_func} returns the levels of a factor

\code{labels_func} returns the label of a variable

\code{observed_func} returns the number of observed values in a variable

\code{missing_func} returns the number of missing values in a variable

\code{distinct_func} returns the number of distinct values in a variable
}
\keyword{internal}
