% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{type_checks}
\alias{type_checks}
\alias{is_unordered}
\alias{is_likert}
\alias{is_character_factor_logical_or_likert}
\alias{is_numeric_logical_or_ordered}
\alias{is_numeric_logical_ordered_or_date}
\title{Type checks}
\usage{
is_unordered(x)

is_likert(x, min_x = 0, max_x = 10)

is_character_factor_logical_or_likert(x)

is_numeric_logical_or_ordered(x)

is_numeric_logical_ordered_or_date(x)
}
\arguments{
\item{x}{a variable}

\item{min_x}{minimum value of a possible Likert scale. Defaults to 0.}

\item{max_x}{maximum value of a possible Likert scale. Defaults to 10.}
}
\value{
a logical
}
\description{
Type checks
}
\details{
\code{is_unordered} checks if an object is an ordered factor

\code{is_likert} checks if an object is a Likert rating scale. If an object has integer values in a limited range it is considered to be a Likert scale.

\code{is_character_factor_logical_or_likert} checks if a vector is a character, factor, a logical or Likert scale.

\code{is_numeric_logical_or_ordered} checks if a vector is numeric, logical or ordered

\code{is_numeric_logical_ordered_or_date} checks if a vector is a numeric, logical, ordered or a date
}
\keyword{internal}
